feature: add clubDetail screen
feature: add clubDetail screen
