+++ feature/club/src/main/java/in/koreatech/koin/feature/club/state/ClubDetailState.kt
@@ -4,6 +4,7 @@ import `in`.koreatech.koin.domain.model.club.ClubDetails
 import `in`.koreatech.koin.domain.model.club.ClubQnasInfo
 
 data class ClubDetailState(
+    val isLoading: Boolean = true,
     val isLogin: Boolean = false,
     val clubDetails: ClubDetails? = null,
     val clubQnasInfo: ClubQnasInfo? = null
