+++ feature/club/src/main/java/in/koreatech/koin/feature/club/type/DetailTabType.kt
@@ -1,0 +1,10 @@
+package `in`.koreatech.koin.feature.club.type
+
+import `in`.koreatech.koin.feature.club.R
+
+enum class DetailTabType(
+    val strResId: Int
+) {
+    DETAIL_INTRO(R.string.club_detail_tab_detail),
+    QNA(R.string.club_detail_tab_qna)
+}
