+++ feature/club/src/main/java/in/koreatech/koin/feature/club/ui/ClubActivity.kt
@@ -6,11 +6,13 @@ import androidx.activity.compose.setContent
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
+import dagger.hilt.android.AndroidEntryPoint
 import `in`.koreatech.koin.core.designsystem.theme.KoinTheme
 import `in`.koreatech.koin.core.designsystem.util.enableEdgeToEdgeWithLightStatusBar
 import `in`.koreatech.koin.feature.club.navigation.ClubNavType
 import `in`.koreatech.koin.feature.club.navigation.koinClubGraph
 
+@AndroidEntryPoint
 class ClubActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
