--- feature/club/src/main/java/in/koreatech/koin/feature/club/ui/detail/type/DetailTabType.kt
@@ -1,8 +1,0 @@
-package `in`.koreatech.koin.feature.club.ui.detail.type
-
-enum class DetailTabType(
-    val title: String
-) {
-    DETAIL_INTRO("상세 소개"),
-    QNA("Q&A")
-}
